requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
PyNaCl==1.5.0
base58==2.1.1
